adjusted delete route

+++ src/main/java/com/luv2code/springboot/cruddemo/rest/EmployeeRestController.java
@@ -56,7 +56,15 @@ public class EmployeeRestController {
     }
 
     @DeleteMapping("/employees/{employeeId}")
-    public void deleteById(@PathVariable int employeeId) {
+    public String deleteById(@PathVariable int employeeId) {
+        Employee employee = employeeService.findById(employeeId);
+
+        if (employee == null) {
+            throw new RuntimeException("Employee id not found - " + employeeId);
+        }
+
         employeeService.deleteById(employeeId);
+
+        return "Deleted employee id - " + employeeId;
     }
 }
